Address PR #2077 review: dispatch gating + reward atomicity
- UpgradeItemHandler: gate on `target is not NpcComponentBundle` so
  players passed as target can't open the upgrade window. Matches
  BazaarHandler / OpenShopHandler / OpenProductionHandler.
- SpChargerHandler: claim the map item via TryRemoveMapItem before
  granting SP, so simultaneous pickups don't both get the reward.
- QuestService: for quests with a q_pay step, CompleteQuestAsync no
  longer publishes QuestCompletedEvent — QPayAsync publishes it when
  CompletedOn is actually set. Previously it fired twice per quest.
- RewardService.DistributeAsync: wrap reward pipeline in try/finally
  so victim.HitList.Clear always runs even if AwardExperienceAsync
  throws.
- UpgradeOperation.ConsumeMaterials: drop redundant (short) cast on
  Math.Min(short, short) (CodeQL finding).
- UpgradeItemHandlerTests: replace the positive-path test (which
  exercised the old overly-permissive null-check behavior) with a
  negative NonNpcTargetIsIgnored spec mirroring TeleporterHandlerTests.

+++ src/NosCore.GameObject/Services/UpgradeService/UpgradeOperation.cs
@@ -203,7 +203,7 @@ public abstract class UpgradeOperation(IRandomNumberSource random, IGameLanguage
                 {
                     break;
                 }
-                var taken = (short)Math.Min(remaining, inst.ItemInstance!.Amount);
+                var taken = Math.Min(remaining, inst.ItemInstance!.Amount);
                 session.Character.InventoryService.RemoveItemAmountFromInventory(taken, inst.ItemInstanceId);
                 remaining -= taken;
                 outPackets.Add(inst.GeneratePocketChange((PocketType)inst.Type, inst.Slot));
